split transform code as standard file
